refactor: 성능 최적화를 위해 이용약관 데이터를 서버에서 받아오도록 리팩토링
